Backoffice - group (#4)
* backoffice - group module readonly

* backoffice - group allow update

* group regular trainings tab

* group regular trainings rework

* remove counts

* remove schemas

* refactor

+++ packages/backend/src/pending-payment/schema.ts
@@ -11,7 +11,7 @@ export const pendingPaymentDetailSchema = z.object({
   connectedPaymentType: z.string(),
   processed: z.boolean(),
   internalNote: z.string(),
-  paymentsCount: z.number().int(),
+  paymentsCount: z.number(),
 });
 
 export type PendingPaymentDetailType = z.infer<
